+++ pyldplayer/__init__.py
@@ -1,7 +1,7 @@
 
 # console
-from pyldplayer.core.console_instance import LDConsoleInstance
-from pyldplayer.core.console_player import LDConsolePlayer
+from pyldplayer.core.instance import LDConsoleInstance
+from pyldplayer.core.player import LDConsolePlayer
 from pyldplayer.core.config import LDPlayerConfig
 
 # mapping
